% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JRRS.R
\name{JRRS}
\alias{JRRS}
\title{linear regression with rank restraint using R.}
\usage{
JRRS(x, y)
}
\arguments{
\item{x}{predictor}

\item{y}{output}
}
\value{
an estimator of regression coefficients\code{n}
}
\description{
linear regression with rank restraint using R
}
\examples{
\dontrun{
JRRS <- function(x,y)
}
}
