% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSC.R
\name{RSC}
\alias{RSC}
\title{linear regression with rank restraint using R.}
\usage{
RSC(x, y)
}
\arguments{
\item{x}{predictor}

\item{y}{output}
}
\value{
the rank of regression coefficients\code{n}
}
\description{
linear regression with rank restraint using R
}
\examples{
\dontrun{
rnR <- gibbsR(100,10)
}
}
