% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{A random walk Metropolis sampler using Rcpp}
\usage{
rwMetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{parameter for produces a normal distribution random sample}

\item{x0}{Initial value}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
sigma <- c(.05, .5, 2, 16)
x0 <- 25
rw1c <- rwMetropolisC(sigma[1], x0, N)
}
}
